listar e cadastrar lancamentos

+++ src/main/java/com/example/algamoney/api/controller/CategoriaController.java
@@ -40,7 +40,7 @@ public class CategoriaController extends GenericController {
 	}
 	
 	@GetMapping("/{id}")
-	public ResponseEntity<?> get(@PathVariable Integer id) {
+	public ResponseEntity<?> get(@PathVariable Long id) {
 		Categoria categoria = categoriaService.obterPorId(id);
 		return ResponseEntity.ok(categoria);
 	}
